[tlv] change GetSize() return type to uint16_t to avoid overflow (#2870)
Credit to OSS-Fuzz.

+++ src/core/common/tlvs.hpp
@@ -104,7 +104,7 @@ public:
      * @returns The total size include Type, Length, and Value fields.
      *
      */
-    uint8_t GetSize(void) const { return sizeof(Tlv) + mLength; }
+    uint16_t GetSize(void) const { return sizeof(Tlv) + mLength; }
 
     /**
      * This method returns a pointer to the Value.
